Time: 203 ms (93.07%), Space: 75.9 MB (41.11%) - LeetHub

+++ 2530-maximal-score-after-applying-k-operations/2530-maximal-score-after-applying-k-operations.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
-    long long maxKelements(vector<int>& nums, int k) {
+    long  maxKelements(vector<int>& nums, int k) {
         priority_queue<int> store;
-        long long ans = 0;
+        long  ans = 0;
         
         for(auto it:nums)
             store.push(it);
         
         
         for(int i = 0; i < k; i++){
-            long long j = store.top();
+            long  j = store.top();
             store.pop();
             ans += j;
             j = ceil(j/3.0);
